# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuquet Automa Core Bridge API",
    "description": "Lightweight native process launcher, WebSocket event hub, and SQLite storage bridge for Automa WebExtension.",
    "contact": {
      "name": "Tuquet Ecosystem",
      "url": "https://github.com/tuquet/tuquet-automa"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8765",
      "description": "Local Daemon Bridge Server"
    }
  ],
  "paths": {
...
    "/api/v1/history": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Get paginated job execution history",
        "description": "Queries previous workflow execution runs and audit logs persisted in the SQLite database.",
        "operationId": "get_job_history",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of history entries to return (default 50)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of items to skip for pagination (default 0)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "example": 0
          },
          {
            "name": "search",
            "in": "query",
            "description": "Optional search query to filter history by workflow name or job ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Optional status filter (e.g. \"completed\", \"failed\", \"running\")",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of past job execution summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobHistoryItem"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database or thread pool failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "History"
        ],
        "summary": "Clear entire job execution history",
        "description": "Permanently truncates the job execution history and purge all logs from the database.",
        "operationId": "clear_all_job_history",
        "responses": {
          "200": {
            "description": "All job history cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryActionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database purge failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{job_id}": {
      "delete": {
        "tags": [
          "History"
        ],
        "summary": "Delete a single job history item",
        "description": "Removes a specific job run and its associated logs from the SQLite history store.",
        "operationId": "delete_job_history_item",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "Unique job identifier to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job history entry deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryActionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database deletion failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{job_id}/logs": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Get execution details and logs for a job",
        "description": "Retrieves full execution audit trail, timing breakdown, and detailed step logs for a specific job.",
        "operationId": "get_job_execution_logs",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "description": "Unique job identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed job execution record and logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetails"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database query failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "status"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "HistoryActionResponse": {
        "type": "object",
        "description": "Response returned after performing a history modification or purge action",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Detailed status message"
          },
          "success": {
            "type": "boolean",
            "description": "Whether the history action succeeded"
          }
        },
        "example": {
          "message": "Job history deleted",
          "success": true
        }
      },
...
      "JobDetails": {
        "type": "object",
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "job": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JobInfo"
              }
            ]
          },
          "logs": {
            "type": [
              "array",
              "null"
            ],
            "items": {}
          },
          "results": {}
        }
      },
      "JobHistoryItem": {
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "JobInfo": {
        "type": "object",
        "required": [
          "id",
          "name",
          "workflowId",
          "status",
          "createdAt",
          "endedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "duration": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "endedAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "workflowId": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}